changed int to float in db and made win/lose conditions

+++ db.py
@@ -2,7 +2,7 @@ def read_money(PLAYER_MONEY):
     try:
         with open(PLAYER_MONEY, "r") as file:
             money = file.read()
-            return int(money)
+            return float(money)
     except FileNotFoundError:
         print("Player money file not found!")
